+++ core/triggers/interface_99_modDoc2Project_doc2Projecttrigger.class.php
@@ -171,11 +171,11 @@ class InterfaceDoc2Projecttrigger
 
 			}
 
+
 		}
 		elseif ($action == 'LINEBILL_INSERT' && $object->product_type != 9 && GETPOST('origin', 'alpha') == 'commande'
 			|| $action == 'LINEBILL_INSERT' && $object->product_type != 9 && GETPOST('origin', 'alpha') == 'propal'
-			|| $action == 'LINEBILL_INSERT' && $object->product_type != 9 && GETPOST('origin', 'alpha') == 'facture'
-		)
+			|| $action == 'LINEBILL_INSERT' && $object->product_type != 9 && GETPOST('origin', 'alpha') == 'facture')
 		{
 			//Récupération des %tages des tâches du projet pour les associer aux lignes de factures
 			$facture = new Facture($db);