+++ ChangeLog.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 ## Unreleased
 
 ## 3.8
+- FIX : FIX COMPAT V22 - *2025-07-07* - 3.8.1
 - NEW : TK2504-3669 Link product to Task thanks to fk_product extrafield - *2025-04-04* - 3.8.0
 
 ## 3.7